refactor(trigger): extract fire paths and run bookkeeping into run.go
Pure move of run execution out of engine.go: the manual fire entry
points (FireManual/FireManualWithActor/FireFromTask), run bookkeeping
(startRun/startRunWithParent, cleanup, WaitRun, KillRun,
finalizeCancelled), the fireKinded/fireAsync/fireSync fire paths,
if_missing prereq resolution, and executor dispatch. engine.go now
holds the Engine struct, constructor/setters, registration routing,
and cron scheduling. No behavior change; code relocated verbatim.

Part of #388 (god-file decomposition).
